+++ src/sora-cpp-sdk/include/sora/hwenc_nvcodec/nvcodec_v4l2_capturer.h
@@ -24,18 +24,15 @@ struct NvCodecV4L2CapturerConfig : V4L2VideoCapturerConfig {
 
 class NvCodecV4L2Capturer : public V4L2VideoCapturer {
  public:
-  static webrtc::scoped_refptr<V4L2VideoCapturer> Create(
+  static webrtc::scoped_refptr<NvCodecV4L2Capturer> Create(
       const NvCodecV4L2CapturerConfig& config);
   NvCodecV4L2Capturer(const NvCodecV4L2CapturerConfig& config);
 
  private:
-  static webrtc::scoped_refptr<V4L2VideoCapturer> Create(
-      webrtc::VideoCaptureModule::DeviceInfo* device_info,
-      const NvCodecV4L2CapturerConfig& config,
-      size_t capture_device_index);
-
+  int32_t Init() override;
   void OnCaptured(uint8_t* data, uint32_t bytesused) override;
 
+  NvCodecV4L2CapturerConfig config_;
   std::shared_ptr<NvCodecDecoderCuda> decoder_;
 };